requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
admin-totals==1.0.1       # via -r requirements.in
boto3==1.13.2             # via -r requirements.in
botocore==1.16.2          # via boto3, s3transfer
brotlipy==0.7.0           # via -r requirements.in
certifi==2020.4.5.1       # via requests, sentry-sdk
cffi==1.14.0              # via brotlipy
chardet==3.0.4            # via requests
django-admin-autocomplete-filter==0.4  # via -r requirements.in
django-admin-rangefilter==0.6.0  # via -r requirements.in
django-crispy-bulma==0.2  # via -r requirements.in
django-crispy-forms==1.7.2  # via -r requirements.in, django-crispy-bulma
django-environ==0.4.5     # via -r requirements.in
django-file-resubmit==0.5.2  # via -r requirements.in
django-recaptcha==2.0.6   # via -r requirements.in
django-spurl==0.6.6       # via -r requirements.in
django-storages==1.9.1    # via -r requirements.in
django==2.2.12            # via -r requirements.in, admin-totals, django-admin-autocomplete-filter, django-crispy-bulma, django-recaptcha, django-storages
docutils==0.15.2          # via botocore
gevent==20.5.0            # via -r requirements.in
greenlet==0.4.15          # via gevent
gunicorn==20.0.4          # via -r requirements.in
idna==2.9                 # via requests
jmespath==0.9.5           # via boto3, botocore
pillow==7.0.0             # via -r requirements.in
psycopg2==2.8.5           # via -r requirements.in
pycparser==2.20           # via cffi
python-dateutil==2.8.1    # via botocore
pytz==2020.1              # via django
requests==2.23.0          # via -r requirements.in
s3transfer==0.3.3         # via boto3
sentry-sdk==0.14.3        # via -r requirements.in
six==1.14.0               # via django-spurl, python-dateutil
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via botocore, requests, sentry-sdk
urlobject==2.4.3          # via django-spurl
whitenoise==5.0.1         # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
